Rename extension to memcached_rb
